requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.23
aiosqlite==0.17.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
requests==2.26.0
starlette==0.27.0
fastapi-mail==1.4.1
pydantic==2.5.2
python-dotenv==1.0.0
email-validator==2.1.0
bcrypt==4.0.1
arxiv==2.0.0
aiosqlite==0.19.0